move bundle classes to change name space
