+++ lib/zakuro/result/operation/month/annotation.rb
@@ -9,11 +9,11 @@ module Zakuro
     #
     # 運用情報
     #
-    module Operation
+    class Operation
       #
       # Month 運用情報（月）
       #
-      module Month
+      class Month
         #
         # Annotation 注釈
         #
